requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0
requests>=2.28.0
pyyaml>=6.0
streamlit>=1.28.0
plotly>=5.18.0
scipy>=1.11.0
tqdm>=4.65.0
pytest>=7.4.0
jupyter>=1.0.0
